refactor: split ordering rules into CodingConventionsTest
Move the 5 Kotlin Coding Conventions class-layout rules out of
ArchitectureTest into a new CodingConventionsTest. ArchitectureTest
now only contains layer-dependency rules.
